uno a molti

+++ app/Http/Controllers/Admin/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Post;
+use App\Category;
 use Illuminate\Support\Str;
 
 class PostController extends Controller
@@ -16,7 +17,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::all();
+        $posts = Post::with('category')->get();
         return view('admin.posts.index', compact('posts'));
     }
 
@@ -27,7 +28,8 @@ class PostController extends Controller
      */
     public function create()
     {
-        return view('admin.posts.create');
+        $categories = Category::all();
+        return view('admin.posts.create', compact('categories'));
     }
 
     /**
@@ -44,6 +46,14 @@ class PostController extends Controller
           ]);
         $dati = $request->all();
         $slug = Str::of($dati['title'])->slug('-');
+        $slug_originale = $slug;
+        $post_trovato = Post::where('slug', $slug)->first();
+        $contatore = 0;
+        while ($post_trovato) {
+            $contatore++;
+            $slug = $slug_originale . '-' . $contatore;
+            $post_trovato = Post::where('slug', $slug)->first();
+        }
         $dati['slug'] = $slug;
         $nuovo_post = new Post();
         $nuovo_post->fill($dati);
@@ -60,7 +70,11 @@ class PostController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
-        return view('admin.posts.show', compact('post'));
+        if ($post) {
+            return view('admin.posts.show', compact('post'));
+        } else {
+            return abort('404');
+        }
     }
 
     /**
@@ -73,7 +87,12 @@ class PostController extends Controller
     {
         $post = Post::find($id);
         if ($post) {
-            return view('admin.posts.edit', compact('post'));
+            $categories = Category::all();
+            $data = [
+                'post' => $post,
+                'categories' => $categories
+            ];
+            return view('admin.posts.edit', $data);
         } else {
             return abort('404');
         }
@@ -95,6 +114,14 @@ class PostController extends Controller
 
         $dati = $request->all();
         $slug = Str::of($dati['title'])->slug('-');
+        $slug_originale = $slug;
+        $post_trovato = Post::where('slug', $slug)->first();
+        $contatore = 0;
+        while ($post_trovato) {
+            $contatore++;
+            $slug = $slug_originale . '-' . $contatore;
+            $post_trovato = Post::where('slug', $slug)->first();
+        }
         $dati['slug'] = $slug;
         $post = Post::find($id);
         $post->update($dati);
@@ -110,7 +137,11 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
-        return redirect()->route('admin.posts.index');
+        if ($post) {
+            $post->delete();
+            return redirect()->route('admin.posts.index');
+        } else {
+            return abort('404');
+        }
     }
 }
